requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
django-dotenv==1.4.2
django-environ==0.11.2
django-environment==0.3.8
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
isort==5.12.0
mysqlclient==2.2.0
pbr==5.11.1
Pillow==10.1.0
platformdirs==3.11.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
stevedore==5.1.0
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
